nvgpu: gpu: Fix misra rule 10.3 in vm unit
For getting mapping kind is passed as signed 32 bit whereas it is stored
as unsigned 32 bit. So, change the kind type to s16 in struct
nvgpu_mapped_buf and also in the declaration from int to s16 to address
that. This is a dependent change for qnx
https://git-master.nvidia.com/r/#/c/2174451/.

Jira NVGPU-3891

Reviewed-on: https://git-master.nvidia.com/r/2176497
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/mm/vm.c
@@ -1252,7 +1252,7 @@ int nvgpu_vm_map(struct vm_gk20a *vm,
 	mapped_buffer->vm           = vm;
 	mapped_buffer->flags        = binfo.flags;
 	nvgpu_assert(map_key_kind >= 0);
-	mapped_buffer->kind         = (u32)map_key_kind;
+	mapped_buffer->kind         = map_key_kind;
 	mapped_buffer->va_allocated = va_allocated;
 	mapped_buffer->vm_area      = vm_area;
